+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 - Renamed `*Plume` classes to `*P`, for brevity; for example, use `CollectionsP` instead of `CollectionsPlume`.
 - Removed all deprecated classes and mehods.
 
+## 1.10.0 (2024-??-??)
+
+- `OrderedPairIterator`: uses `IPair` instead of `MPair`
+- `ArraysPlume`:
+   * deprecated `sorted()` in favor of `isSorted()`
+
 ## 1.9.3 (2024-04-19)
 
 - `CollectionsPlume`: